libmount: Fix uninitialized variable problem
If my reading of the code is correct, then the sysroot_rc and sysroot
members may be used uninitialized. There is no warning for this from
the compiler but that seems to be because the checking for uninitialized
variables does not span over function calls.

This bug was introduced in the recent commit 4e8a8d359
("libmount: Reuse the guessed root device").

Signed-off-by: Viktor Rosendahl (BMW) <viktor.rosendahl@gmail.com>